+++ lib/widgets/journal/entry_details_widget.dart
@@ -23,6 +23,7 @@ import 'package:lotti/widgets/journal/entry_image_widget.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/journal/helpers.dart';
 import 'package:lotti/widgets/journal/journal_card.dart';
+import 'package:lotti/widgets/journal/measurement_summary.dart';
 import 'package:lotti/widgets/journal/tags_widget.dart';
 import 'package:lotti/widgets/misc/survey_summary.dart';
 import 'package:lotti/widgets/tasks/task_form.dart';
@@ -175,7 +176,7 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
                       );
                     },
                   ),
-                  item.maybeMap(
+                  item.map(
                     journalAudio: (JournalAudio audio) {
                       return const AudioPlayerWidget();
                     },
@@ -189,11 +190,12 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
                     survey: (SurveyEntry surveyEntry) =>
                         SurveySummaryWidget(surveyEntry),
                     quantitative: (qe) => Padding(
-                      padding: const EdgeInsets.symmetric(
-                        horizontal: 8.0,
-                        vertical: 16,
-                      ),
+                      padding: const EdgeInsets.all(8),
                       child: InfoText(entryTextForQuant(qe)),
+                    ),
+                    measurement: (measurementEntry) => Padding(
+                      padding: const EdgeInsets.all(8),
+                      child: MeasurementSummary(measurementEntry),
                     ),
                     task: (Task task) {
                       final formKey = GlobalKey<FormBuilderState>();
@@ -245,9 +247,9 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
                         task: task,
                       );
                     },
-                    orElse: () {
-                      return const SizedBox.shrink();
-                    },
+                    habitCompletion: (_) => const SizedBox.shrink(),
+                    journalEntry: (_) => const SizedBox.shrink(),
+                    journalImage: (_) => const SizedBox.shrink(),
                   ),
                   EntryDetailFooter(
                     itemId: widget.itemId,