+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed:
 - Improve layout of health data entry
+- Improve layout of measurable data entry
 
 ## [0.8.45] - 2022-06-12
 ### Fixed: